purrpurragent/requirements.txt
Pillow==10.4.0
google-cloud-aiplatform==1.58.0